frame example codes

+++ byfeature/src/main/kotlin/com/ndgndg91/byfeature/global/protocol/Frame.kt
@@ -1,7 +1,5 @@
 package com.ndgndg91.byfeature.global.protocol
 
-import java.math.BigInteger
-
 data class Frame(
-    val cursor: BigInteger
+    val cursor: Long
 )
